[proliferate] split collapsed action rows
