+++ desktop/src/components/workspace/chat/tool-calls/CollapsedActionRowPrimitives.tsx
@@ -1,0 +1,101 @@
+import { useCallback } from "react";
+import { Button } from "@/components/ui/Button";
+import { ChevronRight } from "@/components/ui/icons";
+import { useOpenInDefaultEditor } from "@/hooks/editor/workflows/use-open-in-default-editor";
+import { useWorkspacePath } from "@/providers/WorkspacePathProvider";
+
+const CHAT_BUTTON_TEXT_CLASS = "text-[length:var(--text-chat)] leading-[var(--text-chat--line-height)]";
+
+export function PlainActionRow({
+  label,
+  tone = "normal",
+}: {
+  label: string;
+  tone?: "normal" | "failed";
+}) {
+  return (
+    <div
+      title={label}
+      className={`truncate text-chat leading-[var(--text-chat--line-height)] ${
+        tone === "failed" ? "text-destructive/80" : "text-muted-foreground/80"
+      }`}
+    >
+      {label}
+    </div>
+  );
+}
+
+export function ActionDisclosureRow({
+  label,
+  expanded,
+  failed,
+  onToggle,
+}: {
+  label: string;
+  expanded: boolean;
+  failed: boolean;
+  onToggle: () => void;
+}) {
+  return (
+    <Button
+      type="button"
+      variant="ghost"
+      size="sm"
+      data-chat-transcript-ignore
+      className={`group/action-row h-auto max-w-full justify-start gap-1 rounded-none bg-transparent p-0 text-left ${CHAT_BUTTON_TEXT_CLASS} font-normal hover:bg-transparent focus-visible:ring-0 ${
+        failed ? "text-destructive/80 hover:text-destructive" : "text-muted-foreground/80 hover:text-foreground"
+      }`}
+      onClick={onToggle}
+    >
+      <span className="min-w-0 truncate">{label}</span>
+      <ChevronRight
+        className={`size-3 shrink-0 text-faint opacity-0 transition-all duration-200 group-hover/action-row:opacity-100 group-focus-visible/action-row:opacity-100 ${
+          expanded ? "rotate-90 opacity-100" : ""
+        }`}
+      />
+    </Button>
+  );
+}
+
+export function ActionFileLink({
+  pathLabel,
+  workspacePath,
+  displayName,
+}: {
+  pathLabel: string;
+  workspacePath: string | null;
+  displayName: string;
+}) {
+  const { resolveAbsolute } = useWorkspacePath();
+  const { openInDefaultEditor } = useOpenInDefaultEditor();
+  const absolute = workspacePath ? resolveAbsolute(workspacePath) : null;
+  const handleOpen = useCallback(() => {
+    if (!absolute) return;
+    void openInDefaultEditor(absolute);
+  }, [absolute, openInDefaultEditor]);
+
+  if (!absolute) {
+    return (
+      <span title={pathLabel} className="min-w-0 truncate text-link-foreground">
+        {displayName}
+      </span>
+    );
+  }
+
+  return (
+    <Button
+      type="button"
+      variant="ghost"
+      size="sm"
+      data-chat-transcript-ignore
+      title={pathLabel}
+      className={`h-auto min-w-0 rounded-none bg-transparent p-0 text-left ${CHAT_BUTTON_TEXT_CLASS} font-normal text-link-foreground hover:bg-transparent hover:underline focus-visible:ring-0 focus-visible:underline`}
+      onClick={(event) => {
+        event.stopPropagation();
+        handleOpen();
+      }}
+    >
+      <span className="min-w-0 truncate">{displayName}</span>
+    </Button>
+  );
+}
